requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pip-tools==7.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyproject_hooks==1.2.0
python-multipart==0.0.17
requests==2.32.3
scipy==1.14.1
setuptools==75.5.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wheel==0.45.0
